"2:45pm. Let me stop at the interlude. The rest of the Densro volume can wait for later. Now I'll deal with chores.
4:10pm. Donw with chores and bath. Usually I take baths after I am done, but I really had to take it right now otherwise I'd feel uncomfortable the entire time while working. It is not like I have a camera supervising my work hours, so I took advantage of that.

This is one of the few luxuries of being my own boss. Now...

I did the hard part. Now I just need to deal with the rest.

...I am thinking. When I am done with the textures I won't actually be done with the desk. I am going to have to redo the UVs so they get packed into their proper slots. I'll have to start thinking about texel densities as well.

But for now it is fine if I just finish the textures. Later on I'll organize them properly. I am not sure how I will deal with them given that I do not have access to advanced UV tools I saw the hard sufrace guy use. I tried looking for UV packmaster, but I could not find a pirate link for it.

Hmmm, I remember than in the PBR guide, it recommended I use flat colors. Right now I am testing out PBR validate and getting all red for some reason.

Nevermind that.

4:35pm. This is some strong thunder. I guess I'll have to run.

...No, it was just once. The rest seems to be the wind. There is some fairly strong wind all of a sudden. Good thing I finished the chores before the rain came along with it.

...No, actually here it is again. This is bad. Let me run.

6:25pm. I am back. That lasted for a while. Before I resume, let me check lunch.

6:45pm. Done with lunch. Let me get on with it. I'll put in a few hours.

Yesterday, a part of the reason it took me so long besides playing with noise is because I was pushing the limits of my understanding. It was like in programming - I was wondering how much I can treat Painter like Designer.

To make the top pattern, I only needed to make 8 layers and position them, but to work through the implications of various approaches took me hours.

Right now, I can easily do the edge wear, but I am going to have to spend some time playing with brushes.

6:50pm. Let me break out the pen. The patterns are one thing, but even at 4k the textures are too small to accurately represent the detals.

7:05pm. done with another part. This is a piece of cake. I'll be done before I know it. I should be able to deal with the whole desk in an hour. Let me see if I get it right.

8:45pm. So much for getting done an hour. I had to put in a bunch of details I forgot by hand.

Just a bit more and I'll leave those big wood panels on the sides for tomorrow. I won't get fancy with those. I'll just reuse the existing wood material, it is not worth it my time to mess with it.

9pm. Agh, I just realized that Pt does not recognize pen pressure for my pen tablet. Nor the extra buttons. Nevermind that for now.

I put in an extra detail, a burnt black mark, on the metal pieces. What a waste of time. I should not have bothered.

I am really annoyed how difficult it is to move around with the pen in substance painter.

9:05pm. Let me close here. I'll finish the desk tomorrow, and post it in the thread. It took way to long for me to get to the point where I can do it, but I am finally (almost) done with it.

I won't put so much effort into texturing the rest of the props, but the desk is where most of the time the MC will be, so I have done it properly.

Faster, stronger, harder! I need to finish the room. I could really push the last mastery challenge to its proper completion, but nothing will stop me with this one. When the room is done I'll move on to doing the rest of the Heaven's Key scenes.

I am not sure how much hope there is for me to get fast with art, but I'll do my best.

What would really make it go zoom is ignoring all this work and just kitbashing things together in Clarisse. Paying for a subscription for one of the 3 fileshosters supporting CGPersia and getting all the models from there would solve most of my environment making problems.

Getting existing stuff an putting it together would really showcase the power of 3d.

If I do everything from the ground up like now, of course it is going to be slow no matter how good I am.

9:20pm. But I need to do it. I need to know the basics of 3d instead of just going straight for the goal.

There will be things I won't be able to find online and dealing with the room challenge will ensure that I am able to deal those.

9:25pm. Now Dendro. Time to clear that backlog.

If it wasn't for the weather, I'd probably have been done completely by now."