"9:40am. Things are going well. Any mail? Nope. I'll just notify the PMer that I emailed him yesterday.
9:50am. I slept well tonight. Let me chill just a bit and then I will get started on the desk piece. I am looking forward to getting it done properly.

10:10am. Let me get started on the desk piece. My comprehension of beveling really went up yesterday.

11am. This is ridiculous. I am getting my ass kicked by these bevels once again. None of the options fix the weird twisting issues.

11:10am. Gahhhhh! I can't figuring out how to do the bevel for that indented part. It is ridiculous.

https://youtu.be/kOteyoLZ2nk
MESHmachine 0.6 - Unf*ck

Let me watch this.

https://youtu.be/kOteyoLZ2nk?t=77

Ah, I see.

https://youtu.be/kOteyoLZ2nk?t=156

Blender has a redo?

11:25pm.

https://youtu.be/iZi5H73_Z6I
MESHmachine 0.8 - Wedge

I have no idea where I am going wrong. Maybe I don't understand bevels after all.

1:40pm. Let me have breakfast. I am still messing with it.

1:55pm.

///

nyaa.si for anime
rargb or 1337x for movies

///

There is some discussion on pirate sites. I never heard of rargb or 1337x.

2:35pm. Done with breakfast. Don't have chores today. Let me chill a bit and then I will resume.

3:15pm. Let me finish the Dungeon Meshi chapter and I will resume. It is a slow day today. I am just moving things around trying to comprehend how those bevels work.

3:25pm. Let me start.

3:45pm. https://blender.stackexchange.com/questions/32675/how-to-subdivide-part-of-mesh-but-not-apply-the-modifier

Let me try that alternative technique. Splitting apart the mesh at the point you down't want it to subdivided is not a bad idea.

But actually, no. This would result in inadequate geometry.

4:45pm. Let me take a break. Absolutely is nothing is going the way I imagine it.

All I wam trying to do is bevel a random hole. It literally has 5 faces.

How did I almost the entire day on this? And days on this previously?

5:20pm. I finally figured out how to do the corners properly. Let me try it again.

5:45pm. I did it! I finally understand how to do that piece perfectly. Let me just try some other things. There is something I need to make sure of.

6:40pm. Right now I am taking a break trying to digest all my insights. Let me finish reading Tower of God and then I will talk about it. As it turns out, I never used MESHmachine for the trouble parts because it was not capable of doing it. Instead I used the % bevel with supporting edges. That turned out to be so powerful. And now that I am actually thinking in this direction, I realize that I could use the technique to make any kind of edge.

7:05pm.  Ok, let me rant a bit. I won't do any art today apart integrating the new piece. The indent is a bit too large so I have to rescale it.

I've actually spent all this time working on the side that won't even be visible in the scene simply because I wanted to know the right way to do it. And now I know.

I was so stupid not because I took the time to learn, but because the techniques I used today are really so obvious yet my head was empty.

Remember how I did it with subdiv by using controlling loop cuts. You can use the exact same technique with bevel with % mode. This would allow you to get any kind of edge variation. Whereas the regular modes are very limited. You can only get a fixed offsets of various kinds.

For whatever reason, I was under the impression that % mode is difficult to use compared to the other modes and never thought more deeply about it. Instead I just bashed my head against the wall.

Had I known what I knew today, I probably would not have bothered buying MESHmachine. What was really digging at me from the inside was the lack of flexibility, and it turns out, the regular bevel tool is the flexible one!

It can quite a lot that MeshMachine's Fuse can't.

Still, this problem here was pretty exceptional. I won't have to dig out the other props and furniture pieces in the room.

But, I can't really call myself a 3d artist if I can't deal with making an indent into a cube and then beveling it so the corners are rounded. You can't make progress by avoiding difficult things. I just had to get a grasp on this very basic thing.

I knew that it should have been possible without resorting to subdiv and now I know how.

7:25pm. Now, even though I put in so much time into texturing it last time, I really will have to retexture it from scratch. The reason - I accidentally got rid of the file with the old object, both the .obj and the .blend version. So I have no choice.

Hmmm, let me try something. How would bevels behave with subdiv style loop cuts? The same as with regular ones assuming the loop slide is turned on. And it really should be with % mode. There is no problem.

8:20pm. I am playing around beveling the cube with supporting loop cuts. Let me watch that radial fillet vid again.

8:25pm. https://youtu.be/7xojWBmfx84?t=149
How to Create Variable Radius Bevel in Blender

Here is him using the bevel modifier while using loop edges to control the sliding. Sigh. This is the peril of going past 6pm.

I am pretty engrossed in studying art though. It is not as mentally strenuous as programming and I am having fun making steady progress in my understanding.

https://youtu.be/7xojWBmfx84?t=332

I am watching this again. The way he did this was in the most inefficient way possible. What he should have done is used the knif tool to carve out the support edges and then just beveled them.

8:40pm. I tried this just now and as expected it works great.

Bevel + knife tool is really a combo to keep in mind.

https://youtu.be/7xojWBmfx84?t=417

Using space to make the edges even is something worth keeping in mind.

8:55pm. Forget it. Let me just put the piece into place and I will close.

9:15pm. It seems the area coverage is actually worse than when it was done automatically. I get only 1.67 while the auto one got 1.82. 83% vs 91%.

9:20pm. Sigh, what a masterpiece. Today I've finally grasped the fundamentals of beveling. Over six months into the journey, but better late than never.

Tomorrow I'll retexture it and move on to the rig and the monitor. The rig will be a bit challeging if I decided to go the boolean approach. If I decide to do the holes in Substance Painter using normals, things will be a lot easier. I'll have to try the both approaches since it is my first time."