"11am. I got up an hour ago and have been playing with beveling the cube. I figured out something interesting. When beveling two adjacent edges in % mode, it matters hugely what the angle between them. It it is 45 degrees or less, it won't slide along the opposing edge.
11:45am. At any rate, enough playing with the damn cube. When you bevel stuff on the surface of the cube, you get some bizzare results. It seems the % mode has a lot of stuff built in to automatically select which edges to bevel. It did not even occur to me to wonder that it is strange that the other side is not being beveled yesterday, but maybe it should have. At any rate, I do not need to understand this. All I need to know is what result I want to get. I can just place the guiding edges after that.

If an edge I want is not being selected for splitting, I can just slide it manually into place before doing the bevel.

The functionality of this is really remarkable

11:55am. Enough of studying this. Maybe I am not really suited to art as I get engrossed in these small matter too easily like in programming. In programming this is good, but I am not sure what to think about this here.

12pm. I checked the emails when I got up, but it was just a bunch of NetMQ issues being auto closed. I completely forgot about those by now. No reply to that sponsorship message yet. Nevermind it.

12:05pm. Let me take a break here.

It seems I am into learning fundamental things, but not so much into making the actual props. I felt yesterday that I could really take off running and close on time today, but spenting a bunch of time playing with the cube threw me off again.

Let me just get breakfast.

Incidentally, Estab Life is the new Milky Holmes. It does not have much in common with it in terms of plot or characters, but the spirit is there. Despite being CGI, if Girls Frontline had its quality, it would have been great."