"1pm. Let me do chores here and then I'll get started on making the monitor.
1:45pm. I need to clear my head. Forget the distractions. For the next couple of hours, just focus on being awesome at 3d art. That is the way to go.

There is no need to think about it too much. Heaven's Key will get made if I can become good enough at art. Otherwise I'll have to get a regular job. It has only been 6.5 months so far. My pace of development has not been bad at all. If I go for a six month stretch where I am not getting better, then I can start to get worried.

2:35pm. Gahhhh, I am running into another unforeseen problem. The subdiv workflow is fine in general, but now that I am trying to make an indent I realized something. It is creating these ridiculous flaps around the edges which creases of 1 do not resolve.

2:45pm. Now I just relized that some geometry to the side is messed up.

Focus me, do not give into despair and scrap everything. You have the unfuck tool for a reason!

2:55pm. No let me start from scratch.

I didn't think thigs would turn out like this.

3pm. Let me take a break.

3:15pm. I am back. When I first started on the monitor, I did a loop cut, and then a bevel, but the % mode for some reason did not work!

Now it works. Why didn't it work earlier?

This is so confusing. I didn't expect to encounter this kind of difficulty.

It seems the path to becoming a skilled 3d artist is a long and winding one.

3:30pm. Let me get into the habit of making copies periodically. The undo will only get me so far.

4:25pm. Let me take a break here. I am still stuck on this.

I need to work less and think more.

4:40pm. This has been happening way too often lately. I run into difficulty and instead of stopping to think, I make an endless amount of experiments to try and wrap my head around it.

Let me make some more.

5:10pm. My whole day has gone kaput because of this. There is one thing I can say with certainty - I do not understand subdivision surfaces at all. What the hell?

Forget modeling.

https://youtu.be/mfp1Z1mBClc
Implementing Catmull-Clark Subdivision - Introduction & Rules

Let me study subdivision for a bit. The way subdivion behaves is too unpredictable, I have zero idea what is going on anymore.

https://youtu.be/mfp1Z1mBClc?t=119
> Real progress is not by speed, but by momentum.

This is a good quote.

https://youtu.be/mfp1Z1mBClc?t=201

Ah, I see. This is one way of converting a ngon to a quad.

6:40pm. I figured it out. I finally know make that dedent. Let me try it again.

8:30pm. I got it perfectly now. It was really simple, but I was too retarded to do the right thing. All I had to do to get the right beveling was to bevel the edges and then join the corners with the other side. It is not complicated at all and yet it had me scratching my head all day. Sigh.

The really embarassing thing is that I was messing with bevels for how long now and I still could not figure out the sequence of actions needed to get the result that I want. I have no excuse.

8:40pm. I'll do beter tomorrow. I really have a grasp on how to use bevels effectively with subdivision surfaces now.

I was in the zone the whole day today, and hopefully tomorrow I'll be able to make significant progress on the prop tomorrow. I tend to do well when I am in the zone programming. In 3d art, I am yet to cross the significant threeshold of experience in order to be able to do things smoothly. But I am really close to cross the threshold. I think it should be smooth sailing from here.

The monitor is unique in being one ove of the very few props that needs subdiv surfaces in the room. Its form has a lot of curviness to it. The rig on the other hand has a lot of decal details which will be a whole different challenge that I will have to tackle. For that one I might have to install Boxcutter, or maybe just Hops will do.

The only other prop that would require the subsurf workflow might be the earphones. Maybe the glass as well, but I can't think of anything else. The radiator? There isn't much.

But there are a lot of props in the room. I will continue training myself on them until I feel satisfied. What I am doing now is no different from a regular artist doing still lives. It is a good way of getting proficiency and confidence in modeling. It might seem trivial, but it is not trivial practice by any means.

Now let me take off for the day. It is time for rest."