"2:45pm. I sure am tired. Let me chill a bit more.
3:10pm. Let me start.

https://www.youtube.com/channel/UCYUWarGVD1rdKDOnOSZbrOQ

Let me just check this guy out.

https://youtu.be/-H1-An1DmLA?t=1
cabochan way blend

What the hell are all these scripts?

...Nevermind, I am not into these tutorials without audio.

3:15pm. Let me play a bit with network and then I will start on the monitor.

3:20pm. To network, it makes a difference whether I have a single curve or one cut in the middle.

3:25pm. https://www.youtube.com/channel/UCGZunNuwUjOY9EWS_-z8trA

Let me watch some of the official Moi 3d vids that are linked in the help.

https://www.youtube.com/watch?v=1B8g42hHJqg
Sunburst

I am mentally prepping myself for the next step. Network is really powerful. I can do a lot with it. Nurbs modeling lacks a bit of flexibility that pixel model does, but I am not sure if the deformation that I have in mind should be intrinsic to it. It is just that Moi does not have them.

3:40pm. https://youtu.be/G8p_ZOvfaso
CrownOfClubs

This one is next.

3:50pm. https://youtu.be/G8p_ZOvfaso?t=271

As I think about this, a thought is starting to surface. Could I use these techniques in Illustrator?

Not really. It has curves, but they are 2d. Not buch I can do about that.

4pm. https://youtu.be/1gcyk2g9deQ
Pod

Let me go for this as well.

https://youtu.be/1gcyk2g9deQ?t=253

Hmm, this is nice. I need to keep this technique of drawing contruction lines in mind.

https://youtu.be/1gcyk2g9deQ?t=920

I hadn't realized that window select is adaptive like this. Nice.

4:20pm. Zbrush is 1.75gb split into 4 parts. RizomUV is 180 only. I'll get it right now.

I might as well get Zbrush while I am at it.

4:30pm. Ok, now, by all means I am supposed to start on the monitor, but I just want to step away from the screen.

The feeling I have right now is that I have it.

Moi 3d essentially resolves all the issues that I've had with Blender's modeling. Even in places where Blender could be better, it is not by much. Moi does not lack power, at worst it might lack a tad ergonomics.

But I do not think even that would be the case.

If I did the monitor using subdiv then yes, I'd have more flexibility to adjust it once it is done, but with the network tool, I could do it in less strokes. And what happened is that I applied the subdiv because of the booleans and had to use the lattice mod anyway. Big whif.

4:30pm. Right now I am reviewing in my head. Adobe Illustrator. I studied it for so much, but now I can barely remember what it was about. Whenever I need to trim a screenshot I just load up CSP and do it there.

It is time to let go of a lot of things.

I've been pushing onwards mixing real practice and learning from others. Moi 3d is so simple that half a week is enough to master it in its entirety. Commpared to any other art programs even 2d ones, it is by far the simplest.

What I will do now is break out my tablet and start modeling. I never expected I would be using it for something like this, but it will serve very well.

I've been running simulations in my head the whole day today.

Is the crop of tools Moi offers me enough or not? Yeah, it is enough to replace all my needs in Blender. I got all those plugins and even paid for MESHmachine, but it turns out I won't need them.

4:50pm. I've made it.

It felt like I've been mid 2/5, but now I am only a few props away from being high 2/5. I just have to practice making them in Moi a bit to internalize the lessons and then that will be there.

Not only that, but I have a foundation in sculpting, texturing and unwrapping now as well. With Clarisse I have an answer on how to put together bigger scenes, on other words layouting. And with the Houdini knowledge that I've acquired, I have an answer for procedural generation too.

By the end of next month, I will be high 2/5 and after that I'll be a step away from the pro level.

The way I will hit the pro level is by mastering all the tools at my disposal and sheding away the unecessary.

Remember what Arrimus said about 2d being faster than 3d.

All this 3d knowledge is essentially scaffolding that will help me later in the move to 2d.

If I can master the 3d tools to a sufficient degree as well as become fast at making 2d, I will have the quality and speed to consider myself 3/5. At that point making my own assets whether it be games or illustrations will cease to be a problem. Remember how annoying it was in 2014 to have finished writing those stories, but not being able to do the covers.

When I am able to do the cover for Heaven's Key, I can consider myself as having accomplished something significant.

It might be fighting a war that is long over, but I do not want to go into the future saying that art is not one of my talents. And if get truck'd and reincarnate back into my schooldays, I'll be sure to get top grades much like in programming. That is the way to think here.

5:10pm. Higher mastery for a higher level of power.

I need to gaze up and stretch my hand out towards the stars.

Model -> texture -> layout.

I think that a lot of the workflow present in 3d also manifests itself in 2d as well. Beginners try to skip them and end up with poor results.

I'll definitely find that out.

When I master my 3d workflow, I'll be able to visualize it and apply it to 2d.

it is similar to how I can imagine code flow in my bed away from the screen.

If you sit down and just start writing code ad-hoc then you are lost. The same goes for drawing.

My conjecture from the beginning was that 3d was good training for 2d, and I am going to get a chance to test that. I can't really imagine somebody being 5/5 at 3d and 1/5 at 2d. The real world does not work that way.

5:15pm. Yeah, I've forgotten that up to now. Although I am doing 3d, what I really wanted deep down was to become good at 2d.

The reason why I went the 3d route is because even though I could have just done 2d directly, trying to do that would be like doing Assembly programming. You can never learn what programming is about by using low level languages, instead you have to go functional and pursue abstraction. This is why despite all the effort I put into Blender, I will move into Moi for prop modeling.

5:25pm. Let me step away from the screen.

My heart was weak and it still is. I've suffered from abandoning my earlier path. But I will endure.

And what once I reach the level of being able to do good art quickly, it will all be worth it. Compared to how much time and effort I put into programming, this here is nothing. I need to believe that I can do it. Deep down, I've started to think that given how much work 3d is, will it really be worth it if I can't do it quickly enough? Sometimes depression can be indiced simply by lack of ambition. It should have been obvious that I would need to master 2d as well. I need to believe that 3d will allow me to do that. I need to believe that being able to walk will allow me to run. Walking will never beat running in terms of speed, but I should not try to skip realms and break my leg in the process.

These reviews and brainstorming sessions are not just to make future plans. They are also to make me realize how much ground I've covered.

A journey of a thousand miles is long, but if you look back and see that you've already 50% done, then why would you stop midway?

Right now, I am done with the hard and tedious parts of art. I've struggled a lot, but from here it will be time for the fun stuff. I have no more need for tutorials. I can just focus on making art and quietly, but surely grind my way to the pro level. From here on out I'll have something to show for my effort."