"8:25pm. I remembered how while Arrimus was doing the blade, he put the rail directly. I realized just now he could have used an isocurve and wanted to try it out. This feature is great.
I really love Moi.

https://www.rhino3d.com/features/quadremesh/

I've seen Rhino being linked to Moi several times. And it seems Rhino has Subdiv modeling tools. Here they are advertizing the quad remesher. It might be worth getting just for that.

Hmmm, it is only 300mb on Persia. Given my interests it might be worth checking out.

https://youtu.be/RIbsNurVdb0
An Introduction to SubD (Subdivision Surface Modelling) in Rhino3d v7

Rhino has spline based Subds!

That could be the best of all worlds.

https://youtu.be/RIbsNurVdb0

Rhino also has deformable Nurbs surfaces.

https://youtu.be/f24n2ijh2Vs
Rhino 7- when to use subd

https://youtu.be/f24n2ijh2Vs?t=178

You can convert subd to NURBs.

https://youtu.be/f24n2ijh2Vs?t=243

Does it not have the equivalent of planar in Moi?

https://youtu.be/f24n2ijh2Vs?t=353
> Don't make yourself insane trying to come up with subd topology in order to get this shape.

9:45pm. https://youtu.be/RIbsNurVdb0?t=2123

This is really interesting.

https://vimeo.com/362114742
QuadRemesh in the Rhino 7 WIP

Here is a vid on the quad remesh in Rhino.

10:10pm. The stuff towards the end where they turn a mesh into a subd object using the quad remesher is interesting.

https://youtu.be/HXBvVhvrQ18
Tsplines to Rhino 7's SubD

Let me watch this as well.

https://youtu.be/suv90ahXa5s
Transitioning from NURBS to T-Splines

https://community.foundry.com/discuss/topic/76334/t-splines-vs-subd
> Please can somebody explain to me what is so great about t-splines (not trolling, just ignorant). What does it offer that subd does not?

I am wondering this as well.

///

While this is true of the T-splines in the papers, the T-splines implementation (in the rhino plugin and fusion 360) is the surface type called T-NURCCs (confusing, I know..) So it's a relative of both catmull-clark subds and nurbs, and can have poles and n-gons.

To answer the original question, I think that there are a few reasons why you might use it. One big reason is that it is a subd-compatible surface that has a strong emphasis on NURBS compatibility. NURBS are useful if you want to do things like booleans, exactly cylindrical screw holes, precise wall thickness when you thicken/offset, and so forth. So if you're looking to manufacture the shapes you're modeling, T-splines can be very helpful there.

There are a few other things that are nice, like being able to have edgeloops that can just T-off, without getting a dimple in your surface like you would with a normal pentagon. Also, you can add more control points to a surface locally (like, insert an edge loop) without changing the smooth shape at all, whereas with normal subds you would either subdivide every face or change the shape of the smooth surface. That is more useful near the end of the modeling process. The tools are also more designed like NURBS tools, with a very strong emphasis on curve-based modeling, lots of ways to use curves that aren't common with subds, like tsPipe.

There are reasons not to use T-splines as well. The surface type is more complex, so it can be more work if you use the advanced features (if you don't, then it's just the same as a Subd or Nurbs). The creasing implementation is problematic at the moment, you often get creases where you don't want them, and changing the radius is not as nice as edge weighting. They might not be integrated into your favorite 3d modeler.. :). And you might not need any of the things they give you over subds - if you're mostly creating renderings rather than manufacturing objects in the real world, subds are simpler.

Please note that I am a bit biased, but I tried to be fair.. :)

///

https://computergraphics.stackexchange.com/questions/8482/problems-with-subdivision-surfaces

https://youtu.be/4nbC6__gZfg
Modeling a tub (extended) with T-Splines for Rhino

I didn't understand the answer. Let me watch this for a bit.

https://youtu.be/ljZCDzWrsS4
How to use T-splines for Rhino

Actually, let me watch this.

> It's not possible any more.

It seems that T splines are patented by Autodesk and can't be used anymore. If I want to give them a try, I should get Fusion 360.

https://youtu.be/IaLwNGPQijs
Fusion 360 VS Rhino, which is better

https://discourse.mcneel.com/t/why-nurbs-and-subd-are-not-back-and-forth-compatible/125621/21?page=2

It seems the t spline patent expires in 2024.

https://youtu.be/IaLwNGPQijs?t=544

InspirationalTuts have really great video scenes.

A lot of good comments on this vid.

https://youtu.be/pKLSTQcFTUI
5 Things For Starting CAD Drafting in Rhino7

https://youtu.be/pKLSTQcFTUI?t=560

The way this works is quite similar to Moi.

https://youtu.be/pKLSTQcFTUI?t=671

I keep reading that NURBS, T splines and Subdiv curves are isomorphic, so a workflow that converts between them should be feasible in Rhino.

https://youtu.be/pKLSTQcFTUI?t=825

These keys for ortho movement, I wonder if they would work in Moi?

Nope.

12pm. Let me close the extended session here. I am a bit excited about finding a way to use the subd workflow together with NURBS. I'll get Rhino tomorrow, but I do not really need it for the sake of dealing with the room. Since Rhino has the quad remesher it should be greatly suitable for making any kinds of props for games. I might want to import in it from Moi. I'll also get Zbrush tomorrow.

Everything is coming together nicely isn't it? The reason why finding this is so nice is because I am already familiar with subd from Blender, so rather than being a new thing, it is a straightforward extension of what I had before."