"9pm. Actually I feel like doing it a bit more.
https://youtu.be/fS8e0x03p_s
[Blender] Modeling Realistic Sci-Fi Helmet w/ Rachel Frick | Part 1: Beginning Modeling

This was posted in that old HS thread on /3/.

https://youtu.be/fS8e0x03p_s?t=908

I am just skimming stuff here. The most interesting part to me is how she is retopoing here. Yeah, to get smooth surfaces, blocking out using the sculpting brushes and then switching gears into this might be a really good idea.

Sculpting brushes are good for sketching things out, and subdiv would be good for edge control once you have the basic shape down. I should consider this kind of thing for the monitor if the quad remesher fails me.

I actually hadn't considered doing things like that as it would be too much work, but it might be less work than the possibilities.

Sculpt and retopo for HS modeling? Why not. Maybe the chads do do it like that.

This way would bring all the styles together. I should go for it.

https://youtu.be/fS8e0x03p_s?t=374
> My way of thinking about sculpting is just as a 2d sketch for my 3d model.
> The sculpt itself took only 40m to make.

I should aspire to this.

https://youtu.be/muHEQu_tNAY?t=17

Doing it like this certainly makes more sense than trying to polish it, at least in Blender. Subdiving would give you those smooth polished surfaces along with the edges on its own. But that is assuming you could get the quad topology right in the first place.

https://youtu.be/muHEQu_tNAY?t=678

Here she is mentioning that sometimes it happens with crease that top faces will pinch over the bottom faces, and suggest to just use edge loops when that happens.

9:45pm. https://youtu.be/5Hl3qI81aTc?list=PL4w6jm6S2lzu4pj4SsPdw62qx6tXqpefE&t=10

There is absolutely no reason why I shouldn't be able to do something like this. It really did take her 40m to model it. This is not a speedup like the radio for example.

https://youtu.be/5Hl3qI81aTc?list=PL4w6jm6S2lzu4pj4SsPdw62qx6tXqpefE

She applied the first level and will got a bit further to add extra details to the model.

10:05pm. Hmmm, right! This is what I did wrong with the monitor. I put the loop edges microscopically close isntead of doing the densible thing of applying the modifier once and refining it.

I should have noted it, but with just a single level of subdiving there was no pinching in those corners. They only became a problem afterwards.

https://youtu.be/5Hl3qI81aTc?list=PL4w6jm6S2lzu4pj4SsPdw62qx6tXqpefE&t=681

She applies it again.

10:10pm. https://youtu.be/5Hl3qI81aTc?list=PL4w6jm6S2lzu4pj4SsPdw62qx6tXqpefE&t=779

This method is quite elegant. I should have watched this video earlier. It would have given me a lot of inspiration.

Watching this is the most ideal way of finishing the month.

https://youtu.be/5Hl3qI81aTc?list=PL4w6jm6S2lzu4pj4SsPdw62qx6tXqpefE&t=829

Hmmm, why is she splitting that edge? It does not seem to be doing anything.

https://youtu.be/5Hl3qI81aTc?list=PL4w6jm6S2lzu4pj4SsPdw62qx6tXqpefE&t=878
> It's crazy just, um, just what a couple of loop edge will do to just strengthen those bevels. Those beveled edges like this. And they always catch the light so nicely. And you'll see when I light this how great those beveled edges catch the light. And you can see just little glints shimmering off of them.

https://youtu.be/5Hl3qI81aTc?list=PL4w6jm6S2lzu4pj4SsPdw62qx6tXqpefE&t=1151

Ah, yes compositing. Something I don't care about. Maybe I'll have to at some point.

10:30pm. Let me close for real. This kind of workflow is a closer match to what I should be doing. Instead of ramming 3 subdiv levels off the bat, I should have gradually increased them during my work on the monitor. Cleaning up after booleans couldn't be helped.

Yeah, what I needed is not Zbrush specifically, though I might very well end up using it. I knew I was missing something in subdiv modeling, it was gradually scaling the complexity threeshold.

Those who live by inspiration, die by inspiration. I missed the right thoughts. Even though I saw Arrimus using subdiv I did not get it until now.

10:40pm. I won't put myself into a situation, where I'd need to immitate Cane's technique. Such a messy way of doing it does not really suit my way of working.

But doing the initial step in Moi/Zbrush and retopoing using the subdiv style would suit me quite well. Whether I use Zbrush or Blender for that part does not matter.

I am going to focus on these thoughts and hold them in my sight all the way to completion.

I am going to clear this mastery challenge. Not only will I clear it, I will aim to do it at an record pace starting from here. Nothing is more amateurish than taking weeks and even months what should be done in a day. I've been instane for the past 7 months and it is time that I establish a regular pace when doing art, much like in programming."