"2:15pm. https://www.youtube.com/watch?v=YbphjXJnuvk Back to Basics: Mesh Deform Modifier
Done with chores. Let me watch some of these.

https://youtu.be/YbphjXJnuvk?t=237

I should figure out what shape keys are.

https://youtu.be/YbphjXJnuvk?t=255

Can you use these to immutably deform the mesh. If so that would be really useful.

https://youtu.be/YbphjXJnuvk?t=312

This is really powerful. I really should get more familiar with the deformers shouldn't I?

It is really hard to do certain things with subdiv. If I start adding more points to get precision that would increase the density of the topology. But with this I could get a lot more mileage out of it without necessarily applying it.

https://youtu.be/B11XBZmIIZc
Learn Shape Keys In Under 5 Mins | Blender

2:30pm. So what happens if you change topology while using shape keys?

2:35pm. Let me get rid of that lattice. I'll have to redo some of the work that I did.

It seems nothing problematic happens when you add geometry. A loop cut just gets interpolated.

2:40pm. Since I am at it, let me check out the other deformers. What about the laplacian deform?

https://youtu.be/UKjuk76Icu8
The Laplacian Smooth Modifier

Let me watch this first.

2:45pm. It seems Youtube has the shorts feature. Never saw this before. It sucks.

https://youtu.be/eozO8MYh-FQ?t=52
Blender Secrets - Laplacian Deform Modifier and Hooks

Hmmm, interesting. Let me play around with this for a bit.

3:05pm. Yeah, it is quite nice. Let me go for some more tutorials on this. What does hook vertex group do given that it already has a system of assigning to them?

https://youtu.be/YB55LsMc7P8
Blender 2.8 Tutorial: Surface Deform Modifier

Let me also watch this.

https://docs.blender.org/manual/en/latest/modeling/modifiers/deform/laplacian_deform.html

The docs mention nothing about hooks. Could I use shape keys for the laplacian deform instead?

https://youtu.be/kbfarm-LSIA
The Blender Hook Modifier EXPLAINED! (When do you use this thing?)

Let me take a look at this.

https://youtu.be/kbfarm-LSIA?t=70

Now I am confused whether it was the hook or the laplacian modifier that was responsible for moving the cube.

https://youtu.be/3brXVB1ewtA?t=82

Smart apply and shift click applies the mods and duplicated the object.

3:55pm. Played around with the laplacian mod a bit.

It is quite nice, but I am going to need the mesh deform in order to affect the subdiv'd vertices. Moving the regular ones around would defeat the point.

Though maybe if I put it after the subdiv modifier the vertex groups would get transfered and still work. It does not matter, the cage is the better solution.

Right, before I resume, let me take a proper break. I want to chill for a bit. After that I'll deal with the monitor.

4:20pm. Let me resume. Time to play with the mesh deformer.

I have that hole on the back of the monitor. The subdiv defines the 4 corners, and it is a real pain in the ass to modify the sides themselves without breaking everything.

Let me finally take care of it properly.

4:30pm. I am confused as to what is going on. Why is it just tugging the edges a bit?

4:55pm. I seems I am still messing with this. I do not get it. It does not behave like the lattice at all.

Ah fuck! I did not apply the scale!

5:05pm. I've tried setting the precision to 10. The quality of the mesh deformer's deformations is low. It is worse than the lattice.

Oh, Blender crashed.

I guess 10 was too much. I'll try 7 next time.

5:10pm. No, 7 is too much as well. I'll try 6.

5:15pm. No, it does not make a difference. Change of plans, I do not trust this as I am getting pinching. The lattice is what gives the best results. I am just going to stack those. I wanted to do that, but Hops refuses to do it automatically so I'll just do it by hand.

Actually what if I duplicated the mesh and applied it as a deformer?

5:30pm. No, the modifier works horrible. It's shit. No wonder I've yet to see it recommended.

5:30pm. Let me take a break here.

6pm. Let me resume. I want to get that monitor out of the way. Let me go all in on lattices. They are great. I have some ideas how to make them work better.

6:35pm. Considering they work based on bounding boxes, it is a waste of time that Blender itself does not do the fitting. Well, all is well that ends well. I finally have the model exported.

6:40pm. But who feels like starting texturing now?

Just one thing - remember that thread I opened yesterday, or was it 2 days ago? I completely forgot to check for replies to the question I posted...it seems I haven't gotten any. That takes care of that.

6:40pm. It is so lame that I spent the entire day today on this monitor. The front was easy, but just as I was afraid, I started fiddling around the back trying to get it to fit just right. I knew it would trigger my compulsive urges to get it right and I was right. This was after I spent the whole night cutting that cube with the knife tool and observing the subdiv patterns.

6:45pm. Instead of the lattice, for more difficult examples, I should just apply the mods and make the fits using the elastic move brush. That would have taken care of the problem in a few minutes. Instead I literally spent hours fiddling around trying to get the topology just right. Know yourself, and know your enemy and you shall not waste hours fiddling around.

6:50pm. I'll close here for the day. This will be on time for once. Tomorrow I will try to fall into a streamlined pace of doing the props and texturing them.

Maybe I'd be faster if I just ground things out instead of constructing an elaborate view of whatever it is that I am attempting first, but that is how I learn, so I can't avoid it. Once I attain expertise, I'll be able to just go forward unperturbed. If I was really good, all the props in my room would be done in a week, two tops.

7:05pm. It would have been faster to just stick with the old asset. Topology asside, the overall shape was there. And the slight shading issues would not be visible. But despite the voice urging me to conserve time, I just had to do it. Looking at the monitor now, and seeing how the quads smoothly flow over the surface without any kind of distortion or pinching makes me pleased. I cannot help, but be satisfied with my owrk, as simple as it is.

Next time, I really will just leave the final touches to the sculpting brushes and not mess with deformers like I have today. They are not worth it.

7:15pm. https://boards.4channel.org/a/thread/237359534/hikaru-no-go-vol18

I read Hikaru no Go a decade and a half ago, but I do not remember this story. I guess I must have missed these extra chapters. I really liked the manga, too bad it got cancelled.

8:15pm. Done with lunch. Let me close here. I am into the Hng reading session. Let me continue having this bit of fun. Tomorrow, I'll get the monitor done for real."